+++ crowbar_framework/spec/lib/crowbar/backup/export_spec.rb
@@ -120,8 +120,8 @@ describe Crowbar::Backup::Export do
 
   context "metadata" do
     it "exports metadata" do
-      allow(Node).to receive(:admin_node).and_return(
-        Node.find_by_name("testing.crowbar.com")
+      allow(ChefNode).to receive(:admin_node).and_return(
+        ChefNode.find_node_by_name("testing.crowbar.com")
       )
       # we skip created_at here as it will always be different
       # so we just compare the classes
