+++ crowbar_framework/spec/models/api/crowbar_spec.rb
@@ -16,7 +16,9 @@ describe Api::Crowbar do
       )
     )
   end
-  let!(:node) { Node.find_by_name("testing.crowbar.com") }
+  let!(:node) do
+    Node.where(name: "testing.crowbar.com").first_or_create(name: "testing.crowbar.com")
+  end
 
   before(:each) do
     allow_any_instance_of(Kernel).to(
@@ -112,7 +114,7 @@ describe Api::Crowbar do
         receive(:find).with("roles:ceph-mon AND ceph_config_environment:*").
         and_return([Node.find_by_name("testing.crowbar.com")])
       )
-      allow_any_instance_of(Node).to(
+      allow_any_instance_of(ChefNode).to(
         receive(:run_ssh_cmd).with("LANG=C ceph health 2>&1").
         and_return(exit_code: 0, stdout: "HEALTH_OK\n", stderr: "")
       )
@@ -126,7 +128,7 @@ describe Api::Crowbar do
         receive(:find).with("roles:ceph-mon AND ceph_config_environment:*").
         and_return([Node.find_by_name("testing.crowbar.com")])
       )
-      allow_any_instance_of(Node).to(
+      allow_any_instance_of(ChefNode).to(
         receive(:run_ssh_cmd).with("LANG=C ceph health 2>&1").
         and_return(exit_code: 1, stdout: "HEALTH_ERR\n", stderr: "")
       )
@@ -138,7 +140,7 @@ describe Api::Crowbar do
         receive(:find).with("roles:ceph-mon AND ceph_config_environment:*").
         and_return([Node.find_by_name("testing.crowbar.com")])
       )
-      allow_any_instance_of(Node).to(
+      allow_any_instance_of(ChefNode).to(
         receive(:run_ssh_cmd).with("LANG=C ceph health 2>&1").
         and_return(exit_code: 0, stdout: "HEALTH_WARN", stderr: "")
       )
