+++ crowbar_framework/spec/models/crowbar_service_spec.rb
@@ -21,9 +21,11 @@ describe CrowbarService do
   before do
     allow_any_instance_of(CrowbarService).to receive(:system).and_return(false)
     allow_any_instance_of(CrowbarService).to receive(:run_remote_chef_client).and_return(0)
+    Node.where(name: "testing.crowbar.com").first_or_create(name: "testing.crowbar.com")
   end
 
   let(:crowbar) { c = CrowbarService.new(Logger.new("/dev/null")); c.bc_name = "crowbar"; c }
+  let(:node) { Node.find_by_name("testing.crowbar.com") }
 
   describe "transition" do
     it "returns 404 without state" do
@@ -46,25 +48,26 @@ describe CrowbarService do
     describe "to another state" do
       before do
         @node = Node.find_by_name("testing.crowbar.com")
+        @chef_node = ChefNode.find_node_by_name("testing.crowbar.com")
       end
 
       it "sets the state debug and state" do
-        allow(Node).to receive(:find_by_name).and_return(@node)
+        allow_any_instance_of(Node).to receive(:chef_node).and_return(@chef_node)
         crowbar.transition("default", @node.name, "a state")
         expect(@node.state).to be == "a state"
         expect(@node.crowbar["crowbar"]["state_debug"]).to_not be_empty
       end
 
       it "saves the node" do
-        allow(Node).to receive(:find_by_name).and_return(@node)
-        expect(@node).to receive(:save).at_least(:once)
+        allow_any_instance_of(Node).to receive(:chef_node).and_return(@chef_node)
+        expect_any_instance_of(Node).to receive(:save).at_least(:once)
         crowbar.transition("default", @node.name, "a state")
       end
     end
 
     describe "to testing" do
       it "creates new node if not found" do
-        expect(Node).to receive(:create_new).with("missing").and_return(nil).at_least(:once)
+        expect(Node).to receive(:find_or_create_by).with(name: "missing").at_least(:once)
         crowbar.transition("default", "missing", "testing")
       end
     end
@@ -76,14 +79,12 @@ describe CrowbarService do
 
       it "adds role to the node if admin" do
         expect(crowbar).to receive(:add_role_to_instance_and_node).at_least(:once)
-        crowbar.transition("default", "admin", "discovering")
+        crowbar.transition("default", "admin.crowbar.com", "discovering")
       end
 
       it "creates new node if not found" do
-        allow(Node).to receive(:find_by_name).and_return(nil)
-        expect(Node).to receive(:create_new).and_return(nil).at_least(:once)
+        expect(Node).to receive(:find_or_create_by).with(name: "admin").at_least(:once)
         crowbar.transition("default", "admin", "discovering")
-        crowbar.transition("default", "missing", "discovering")
       end
 
       it "check that the node is initially not allocated" do
@@ -94,38 +95,32 @@ describe CrowbarService do
     end
 
     describe "to hardware-installing" do
-      before do
-        @node = Node.find_by_name("testing.crowbar.com")
-      end
-
       it "forces nodes transition to a given state" do
-        allow(Node).to receive(:find_by_name).and_return(@node)
+        allow(ChefNode).to receive(:find_node_by_name).and_return(
+          ChefNode.find_node_by_name("testing.crowbar.com")
+        )
         crowbar.transition("default", "testing.crowbar.com", "hardware-installing")
-        expect(@node.state).to be == "hardware-installing"
+        expect(node.state).to be == "hardware-installing"
       end
     end
 
     describe "to hardware-updating" do
-      before do
-        @node = Node.find_by_name("testing.crowbar.com")
-      end
-
       it "forces nodes transition to a given state" do
-        allow(Node).to receive(:find_by_name).and_return(@node)
+        allow(ChefNode).to receive(:find_node_by_name).and_return(
+          ChefNode.find_node_by_name("testing.crowbar.com")
+        )
         crowbar.transition("default", "testing.crowbar.com", "hardware-updating")
-        expect(@node.state).to be == "hardware-updating"
+        expect(node.state).to be == "hardware-updating"
       end
     end
 
     describe "to update" do
-      before do
-        @node = Node.find_by_name("testing.crowbar.com")
-      end
-
       it "forces nodes transition to a given state" do
-        allow(Node).to receive(:find_by_name).and_return(@node)
+        allow(ChefNode).to receive(:find_node_by_name).and_return(
+          ChefNode.find_node_by_name("testing.crowbar.com")
+        )
         crowbar.transition("default", "testing.crowbar.com", "update")
-        expect(@node.state).to be == "update"
+        expect(node.state).to be == "update"
       end
     end
   end
